requirements.txt
yfinance
beautifulsoup4
requests
feedparser
google-generativeai
matplotlib
mplfinance
python-dotenv
